chore: eslint packages
- [x] Create base eslint config
- [x] Create react eslint config
- [x] Add tsconfig.json to root
