+++ packages/app/.eslintignore
@@ -1,8 +1,4 @@
 dist/
 node_modules/
 coverage/
-zkeyFiles/
-test-results/
-playwright-report/
-playwright/.cache/
 tmp/
