+++ packages/app/package.json
@@ -43,7 +43,7 @@
   },
   "homepage": "https://github.com/privacy-scaling-explorations/crypt-keeper-extension",
   "devDependencies": {
-    "@babel/eslint-parser": "^7.22.5",
+    "@cryptkeeper/eslint-config-react": "workspace:^",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
     "@types/chrome": "^0.0.237",
@@ -61,8 +61,6 @@
     "@types/redux-logger": "^3.0.9",
     "@types/testing-library__jest-dom": "^5.14.6",
     "@types/webextension-polyfill": "^0.10.0",
-    "@typescript-eslint/eslint-plugin": "^5.60.0",
-    "@typescript-eslint/parser": "^5.60.0",
     "browserify": "^17.0.0",
     "copy-webpack-plugin": "^11.0.0",
     "cors": "^2.8.5",
@@ -71,16 +69,6 @@
     "dotenv": "^16.3.1",
     "env-cmd": "^10.1.0",
     "eslint": "^8.43.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-import-resolver-typescript": "^3.5.5",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-json": "^3.1.0",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.32.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-unused-imports": "^2.0.0",
     "express": "^4.18.2",
     "fast-check": "^3.10.0",
     "file-loader": "^6.2.0",
