+++ packages/e2e/.eslintignore
@@ -1,7 +1,4 @@
-dist/
 node_modules/
-coverage/
-zkeyFiles/
 test-results/
 playwright-report/
 playwright/.cache/
