+++ packages/e2e/package.json
@@ -23,19 +23,13 @@
   "homepage": "https://github.com/privacy-scaling-explorations/crypt-keeper-extension",
   "devDependencies": {
     "@babel/eslint-parser": "^7.22.5",
+    "@cryptkeeper/eslint-config-base": "workspace:^",
     "@playwright/test": "^1.35.1",
     "@synthetixio/synpress": "^3.7.1",
     "@typescript-eslint/eslint-plugin": "^5.60.0",
     "@typescript-eslint/parser": "^5.60.0",
     "eslint": "^8.43.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-import-resolver-typescript": "^3.5.5",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-json": "^3.1.0",
-    "eslint-plugin-playwright": "^0.15.1",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-unused-imports": "^2.0.0",
+    "eslint-plugin-playwright": "^0.14.3",
     "lint-staged": "^13.2.2",
     "prettier": "^2.8.8",
     "pretty-quick": "^3.1.3",
