+++ packages/eslint-config-react/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@cryptkeeper/eslint-config-react",
+  "version": "0.1.0",
+  "description": "ESLint react config for CryptKeeper extension",
+  "private": true,
+  "main": "index.js",
+  "author": "Privacy and Scaling explorations team",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/privacy-scaling-explorations/crypt-keeper-extension"
+  },
+  "homepage": "https://github.com/privacy-scaling-explorations/crypt-keeper-extension",
+  "dependencies": {
+    "@babel/eslint-parser": "^7.22.5",
+    "@cryptkeeper/eslint-config-base": "workspace:^",
+    "@typescript-eslint/eslint-plugin": "^5.60.0",
+    "@typescript-eslint/parser": "^5.60.0",
+    "eslint": "^8.43.0",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0"
+  }
+}
